Moved MapboxDataProperty to separate file; corrected namespaces.
