+++ sdkproject/Assets/Mapbox/Unity/DataContainers/MapboxDataProperty.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: d7e99aece91ce4463b09850af8e3df52
+timeCreated: 1536268482
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
