+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Modifiers/ModifierBase.cs
@@ -3,34 +3,7 @@ namespace Mapbox.Unity.MeshGeneration.Modifiers
 	using UnityEngine;
 	using System;
 	using Mapbox.Unity.MeshGeneration.Data;
-
-
-	public abstract class MapboxDataProperty
-	{
-		public event System.EventHandler PropertyHasChanged;
-		protected virtual void OnPropertyHasChanged(System.EventArgs e)
-		{
-			System.EventHandler handler = PropertyHasChanged;
-			if (handler != null)
-			{
-				handler(this, e);
-			}
-		}
-		bool _hasChanged;
-
-		public bool HasChanged
-		{
-			set
-			{
-				if (value == true)
-				{
-					OnPropertyHasChanged(null /*Pass args here */);
-					// reset HasChanged 
-					_hasChanged = false;
-				}
-			}
-		}
-	}
+	using Mapbox.Unity.Map;
 
 	[Serializable]
 	public abstract class ModifierProperties : MapboxDataProperty
